Move most 'is' checks back to IndividualUser
