+++ src/main/kotlin/com/terraformation/backend/customer/model/DeviceManagerUser.kt
@@ -45,8 +45,6 @@ data class DeviceManagerUser(
     mapOf(facilityId to Role.Contributor)
   }
 
-  override var isRecordingChecks = false
-
   override val defaultPermission: Boolean
     get() = false
