+++ src/main/kotlin/com/terraformation/backend/customer/model/FunderUser.kt
@@ -24,8 +24,6 @@ data class FunderUser(
   override val userType: UserType
     get() = UserType.Funder
 
-  override var isRecordingChecks = false
-
   override fun getName(): String = authId ?: throw IllegalStateException("User is unregistered")
 
   override fun getUsername(): String = authId ?: throw IllegalStateException("User is unregistered")
