+++ src/main/kotlin/com/terraformation/backend/customer/model/SystemUser.kt
@@ -68,8 +68,6 @@ class SystemUser(
   override val email: String?
     get() = null
 
-  override var isRecordingChecks = false
-
   /*
    * The system user has no roles per se; it always has access to everything. Reject any attempts to
    * walk the current user's list of roles.
